requirements.txt
google-cloud-speech
google-cloud-texttospeech
google-generativeai
Pillow
pvporcupine
pyaudio
pydub
python-dotenv
rich
webrtcvad
boto3
requests
azure-cognitiveservices-speech
openai
numpy